refactor(layout): reorganize _src into gaussx-style themed subpackages
Drop the flat _src/ namespace in favour of the layout gaussx uses —
private themed subpackages with the public API re-exported untouched:

- _primitives/  statistics, gain, likelihood, perturbations,
                localization, inflation, inflators
- _filters/     sequential (L1), models (L2), latent, parametric
- _processes/   processes (L1), process_models (L2), schedulers, optax
- _train/       differentiable, road
- _smoothers.py, _diagnostics.py, _types.py, _checks.py, _protocols.py
  stay top-level (single-module concerns / cross-cutting)

Pure move + import-path rewrite; no behaviour change. The public
namespaces (filterax.filters, .smoothers, .processes, .optax,
.differentiable, .utils) and every top-level export are unchanged.
