+++ src/filterax/_checks.py
