+++ src/filterax/_diagnostics.py
