+++ src/filterax/_filters/__init__.py
@@ -1,0 +1,1 @@
+"""Sequential filters: L1 analysis steps and L2 assimilation models."""
