+++ src/filterax/_primitives/__init__.py
@@ -1,0 +1,1 @@
+"""Ensemble-DA primitives: statistics, gain, likelihood, inflation, localization."""
