+++ src/filterax/_primitives/_gain.py
@@ -6,7 +6,7 @@ import gaussx
 import lineax as lx
 from jaxtyping import Array, Float
 
-from filterax._src._checks import check_ensemble_size
+from filterax._checks import check_ensemble_size
 
 
 def kalman_gain(
