+++ src/filterax/_processes/__init__.py
@@ -1,0 +1,1 @@
+"""Ensemble Kalman processes (EKI/EKS/UKI), schedulers, and optax wrappers."""
