--- src/filterax/_src/__init__.py
@@ -1,5 +1,0 @@
-"""Internal implementation namespace.
-
-Public symbols are re-exported from :mod:`filterax`. Import from here only if
-you explicitly need an internal name.
-"""
