+++ src/filterax/_train/__init__.py
@@ -1,0 +1,1 @@
+"""Differentiable assimilation and gradient-based training loops."""
