+++ src/filterax/_train/_differentiable.py
@@ -39,13 +39,13 @@ import jax.numpy as jnp
 import lineax as lx
 from jaxtyping import Array, Float
 
-from filterax._src._protocols import (
+from filterax._protocols import (
     AbstractDynamics,
     AbstractInflator,
     AbstractObsOperator,
     AbstractSequentialFilter,
 )
-from filterax._src._types import AssimilationResult
+from filterax._types import AssimilationResult
 
 
 def _forecast(
@@ -65,10 +65,10 @@ def _reject_stochastic_components(
     """Refuse filter / inflator flavours whose gradients are non-smooth.
 
     Deferred imports avoid a module-level cycle with
-    :mod:`filterax._src.sequential` and :mod:`filterax._src.inflators`.
+    :mod:`filterax._filters._sequential` and :mod:`filterax._primitives._inflators`.
     """
-    from filterax._src.inflators import AdditiveInflator
-    from filterax._src.sequential import ETKF_Livings, StochasticEnKF
+    from filterax._filters._sequential import ETKF_Livings, StochasticEnKF
+    from filterax._primitives._inflators import AdditiveInflator
 
     if isinstance(filter_, (StochasticEnKF, ETKF_Livings)):
         raise ValueError(
