+++ src/filterax/processes/__init__.py
@@ -1,14 +1,14 @@
 """Layer-1 ensemble Kalman process components.
 
 Re-exports the L1 process classes from
-:mod:`filterax._src.processes` so the canonical
+:mod:`filterax._processes._processes` so the canonical
 ``filterax.processes.EKI()`` access path matches the design docs.
 The Layer-2 ``filterax.EKI`` / ``filterax.EKS`` / ``filterax.UKI``
 wrappers live at the top level and own the full
 ``init → forward eval → update`` loop.
 """
 
-from filterax._src.processes import (
+from filterax._processes._processes import (
     EKI as EKI,
     ETKI as ETKI,
     GNKI as GNKI,
